# src/main/java/com/example/bankcards/dto/TransferCardDto.java
package com.example.bankcards.dto;

import java.math.BigDecimal;

public class TransferCardDto {
   private Long fromCardId;
   private Long toCardId;
   private BigDecimal amount;

    public Long getFromCardId() {
        return fromCardId;
    }

    public void setFromCardId(Long fromCardId) {
        this.fromCardId = fromCardId;
    }

    public Long getToCardId() {
        return toCardId;
    }

    public void setToCardId(Long toCardId) {
        this.toCardId = toCardId;
    }

    public BigDecimal getAmount() {
        return amount;
    }

    public void setAmount(BigDecimal amount) {
        this.amount = amount;
    }
}